untested Go rework
