+++ go/pkg/lib/grpc_server.go
@@ -2,6 +2,7 @@ package lib
 
 import (
 	"context"
+	"fmt"
 
 	pb "github.com/Jorrit05/DYNAMOS/pkg/proto"
 	"google.golang.org/protobuf/types/known/emptypb"
@@ -11,7 +12,8 @@ type SharedServer struct {
 	pb.UnimplementedMicroserviceServer
 	pb.UnimplementedHealthServer
 	pb.UnimplementedGenericServer
-	callbacks map[string]func(ctx context.Context, data *pb.MicroserviceCommunication) (*emptypb.Empty, error)
+	ServiceName string
+	Callback    func(ctx context.Context, data *pb.MicroserviceCommunication) error
 }
 
 func (s *SharedServer) Check(ctx context.Context, in *pb.HealthCheckRequest) (*pb.HealthCheckResponse, error) {
@@ -28,36 +30,33 @@ func (s *SharedServer) InitTracer(ctx context.Context, in *pb.ServiceName) (*emp
 	return &emptypb.Empty{}, nil
 }
 
-func (s *SharedServer) RegisterCallback(msgType string, callback func(ctx context.Context, data *pb.MicroserviceCommunication) (*emptypb.Empty, error)) {
-	if s.callbacks == nil {
-		s.callbacks = make(map[string]func(ctx context.Context, data *pb.MicroserviceCommunication) (*emptypb.Empty, error))
-	}
-	s.callbacks[msgType] = callback
-}
-
+// Shared implementation for SendData, this is called by the previous microservice
+// or by the sidecar itself if a message is received from rabbitMQ
+//
+// Parameters:
+//   - ctx: The context of the request
+//   - data: MicroserviceCommunication messages
+//
+// Returns:
+//   - ContinueReceiving: A boolean indicating if the sidecar should continue receiving messages
+//   - error: An error if the function fails
 func (s *SharedServer) SendData(ctx context.Context, data *pb.MicroserviceCommunication) (*pb.ContinueReceiving, error) {
 	logger.Sugar().Debugf("Starting lib.SendData: %v", data.RequestMetadata.DestinationQueue)
 
-	ctx, span, err := StartRemoteParentSpan(ctx, "sidecar SendData/func:", data.Traces)
+	ctx, span, err := StartRemoteParentSpan(ctx, fmt.Sprintf("%s SendData/func:", s.ServiceName), data.Traces)
 	if err != nil {
 		logger.Sugar().Warnf("Error starting span: %v", err)
 	}
 	defer span.End()
 
-	// This callbackmechanism is outdated and can be removed, for this the consumption service of microservices
-	// should be updated and moved to a generic lib.
-	// The MS will process the message and send it to the next MS (or sidecar).
-	callback, ok := s.callbacks[data.Type]
-	if !ok {
-		logger.Warn("no callback registered for this message type")
-
-		return &pb.ContinueReceiving{ContinueReceiving: false}, nil
+	if data.Type == "microserviceCommunication" {
+		err = s.Callback(ctx, data)
+		if err != nil {
+			logger.Sugar().Errorf("Failed to process message: %v", err)
+		}
+	} else {
+		logger.Sugar().Errorf("Unknown message type: %v", data.Type)
+		return &pb.ContinueReceiving{ContinueReceiving: false}, fmt.Errorf("unknown message type: %s", data.Type)
 	}
-
-	if _, err := callback(ctx, data); err != nil {
-		logger.Sugar().Errorf("Callback Error: %v", err)
-		return &pb.ContinueReceiving{ContinueReceiving: false}, nil
-	}
-
-	return &pb.ContinueReceiving{ContinueReceiving: false}, nil
+	return &pb.ContinueReceiving{ContinueReceiving: false}, err
 }
